fix(inference): preserve optional-chain short-circuit types
